fix(gateway): move S3 lifecycle enforcement to provisioning
